requirements.txt
requests
PyJWT
protobuf
protobuf-decoder
pycryptodome
psutil
rich
python-cfonts
urllib3
pyTelegramBotAPI
Flask
